requirements.txt
PyQt6==6.6.1
google-cloud-build==3.15.0
google-cloud-run==0.10.0
google-cloud-artifact-registry==1.8.0
google-cloud-logging==3.8.0
google-cloud-iam==2.12.0
google-cloud-resource-manager==1.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1 